more example renaming
