+++ landing-strip/demo/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
     android:label="@string/app_name"
     android:theme="@style/AppTheme">
 
-    <activity android:name="com.novoda.landingstrip.BasicUsageActivity">
+    <activity android:name="com.novoda.landingstrip.SimpleTextTabActivity">
       <intent-filter>
         <action android:name="android.intent.action.MAIN" />
         <category android:name="android.intent.category.LAUNCHER" />
