+++ landing-strip/demo/src/main/java/com/novoda/landingstrip/CustomLandingStrip.java
@@ -5,6 +5,7 @@ import android.support.v4.view.ViewPager;
 import android.support.v7.app.ActionBarActivity;
 
 import com.novoda.landing_strip.R;
+import com.novoda.landingstrip.setup.DemoPagerAdapter;
 
 public class CustomLandingStrip extends ActionBarActivity {
