+++ landing-strip/demo/src/main/java/com/novoda/landingstrip/setup/Data.java
@@ -1,4 +1,4 @@
-package com.novoda.landingstrip;
+package com.novoda.landingstrip.setup;
 
 public enum Data {
