+++ landing-strip/demo/src/main/res/layout/activity_basic_usage.xml
@@ -10,7 +10,7 @@
     android:layout_height="50dp"
     android:layout_width="match_parent"
     android:background="@android:color/holo_orange_dark"
-    app:tabLayoutId="@layout/tab"
+    app:tabLayoutId="@layout/tab_simple_text"
     app:indicatorColour="@android:color/white" />
 
   <android.support.v4.view.ViewPager
