+++ landing-strip/demo/src/main/res/layout/activity_custom_landing_strip.xml
@@ -14,7 +14,7 @@
       android:id="@+id/landing_strip"
       android:layout_height="50dp"
       android:layout_width="wrap_content"
-      app:tabLayoutId="@layout/tab"
+      app:tabLayoutId="@layout/tab_simple_text"
       app:indicatorColour="@android:color/white" />
 
   </FrameLayout>
