+++ landing-strip/demo/src/main/res/layout/tab_custom.xml
