+++ landing-strip/demo/src/main/res/layout/tab_simple_text.xml
